add JSVal::UInt16

+++ src/lv5/runtime_string.h
@@ -522,8 +522,7 @@ inline JSVal StringCharCodeAt(const Arguments& args, Error* error) {
   if (position < 0 || position >= str->size()) {
     return JSNaN;
   } else {
-    return JSVal::UInt32(
-        static_cast<uint32_t>(str->GetAt(core::DoubleToUInt32(position))));
+    return JSVal::UInt16(str->GetAt(core::DoubleToUInt32(position)));
   }
 }
